requirements.txt
django
django-role-permissions